Implementing Schedule to delete data

+++ src/main/java/com/plataform/courses/services/UserService.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.plataform.courses.model.entity.User;
 import com.plataform.courses.repository.UserRepository;
 import com.plataform.courses.services.exceptions.BadWordException;
+import com.plataform.courses.services.exceptions.NotPermissionImmutableData;
 import com.plataform.courses.services.exceptions.ObjectNotFoundException;
 
 @Service
@@ -19,6 +20,12 @@ public class UserService {
     private UserRepository userRepository;
 
     private static String UNTANTED_CONTENT = "Conteúdo indesejado detectado";
+
+    private static String NOT_PERMISSION_DELETE = "Você não tem permissão para deletar este usuário";
+
+    private static String NOT_PERMISSION_UPDATE = "Você não tem permissão para alterar este usuário";
+
+    private static final Integer MAX_IMMUTABLE_RECORDS = 3;
 
     public User findById(Long id){
         Optional<User> user = this.userRepository.findById(id);
@@ -36,6 +43,12 @@ public class UserService {
                 UNTANTED_CONTENT
             );
         }
+        Long immutableCount = userRepository.countByImmutableTrue();
+        if (immutableCount >= MAX_IMMUTABLE_RECORDS){
+            obj.setImmutable(false);
+        } else {
+            obj.setImmutable(true);
+        }
         obj.setId(null);
         obj.setImmutable(false);
         obj = this.userRepository.save(obj);
@@ -50,6 +63,9 @@ public class UserService {
             throw new BadWordException(UNTANTED_CONTENT);
         }
         User newObj = findById(obj.getId());
+        if (newObj.getImmutable().equals(true)){
+            throw new NotPermissionImmutableData(NOT_PERMISSION_UPDATE);
+        }
         newObj.setName(obj.getName());
         newObj.setEmail(obj.getEmail());
         newObj.setImmutable(false);
@@ -57,7 +73,10 @@ public class UserService {
     }
 
     public void delete(Long id){
-        findById(id);
+        User obj = findById(id);
+        if (obj.getImmutable().equals(true)){
+            throw new NotPermissionImmutableData(NOT_PERMISSION_DELETE);
+        }
         try{
             this.userRepository.deleteById(id);
         } catch (Exception e){
